Remove BAILOUT macro.
There was one place where it was not used correctly, and the entity pointer
is not needed anymore after PCI resources have been claimed. In future, other
things will have to be handled when bailing out of PreInit later on, things
which need more code than a macro should have. If the extra work involved is
that much of a burden, then this is about the only case where a goto is
accepted, especially by the otherwise very picky kernel community.