Track option set status.
Presently there is no consistent way of tracking if an option
was set by the user or has been set to the default value.
This information may be needed later to make more educated
decisions on how to proceed.
Many drivers work around this problem in different ways - often
by introducing a second driver private variable to track the
option status.
We add a set of functions and a new structure to handle this
in a more consistent way.
These functions should eventually be moved to DDX.
For now we leave them in the driver.