Fixed bailout structure: uninit bios data and unmap mmio when needed.
The failure case was still not handled correctly:
MMIO should get unmapped and the BIOS handler private data
structures should be freed on failure. (If the latter ones
will be needed in ScreenInit() in the future we can keep
them around and point to them from the driver private data
structure.