Single entry point to ATOM BIOS subsystem.
The entry point to the ATOM BIOS subsystem is now
a single function:
AtomBiosResult RhdAtomBIOSFunc(ScrnInfoPtr pScrn, pointer handle, AtomBiosFunc func,
		    AtomBIOSArgPtr data);
Apart from pScrn the arguments are:
 'handle' (a struct with persistent internal data),
typedef enum {
    ATOMBIOS_INIT,
    ATOMBIOS_UNINIT,
    ...
    FUNC_END
} AtomBiosFunc;
which describes the action to perform.
Before any other action can be performed ATOMBIOS_INIT needs
to be called. It will create and return the 'handle' which
needs to be passed for any subsequent operations.

it can return the states:
typedef enum {
    SUCCESS,
    FAILED,
    NOT_IMPLEMENTED
} AtomBiosResult;

The 4th argument is a pointer to a union
that can be used to pass arguments to and
results from the function.
The most common data types are already in the union,
more elaborate structures can be passed back and
forth by using the ptr member.

AtomBiosArg