Get videoRam size from MMIO 0x00F8.
This is where Atom stores the size. This size is checked against the size of
the PCI BAR aperture (which is limited to 512MB), and the smallest of the
two is chosen.
Also removes the superfluous BAR selection crap, uses 2 defines instead. Will
reintroduce logic when it is actually needed.