Add one or two pci ids...
We had a delivery of AMD hardware, so we had a bit of a plugfest. This means
that we now have a grand total of 15 unique devices at our disposal.

Then boiko@mandriva dug up a pci id listing from ATI, which kind of bumped
the number of possible R5xx/R6xx devices up to a cosy 99.

We need to order more hardware.

So:
* Add to RHD_CHIPSETS, RHDChipsets, and RHDPCIchipsets.
* Create a function to mark the "few" cards we can't test as experimental,
  which means that the driver will complain and bail.
* Add Option "EXPERIMENTAL", to stop the driver from bailing.
* Add PCI card/subsystem id checking; only prints the card name now.
* Remove rhd_macro.h as RHDPTR() should live near the definition of RHDPtr.
* Move a few other defines, like RHD_NAME, to rhd.h.