Add connector handling.
Connectors tie the actual plug, a monitor of any description, a ddc bus
a hot plug detect pin, and (currently) up to two outputs together.

So, everything comes together here, and the connector handling decides
how and when which monitors and which outputs and which CRTCs are tied
together.

This requires an entry in our card table to map out these connectors
for various hardware, another lengthy and painful plugfest...
Which isn't complete yet, so your hardware might not start just yet.