Don't validate the duty cycle on fixed device provided modes.
Modes provided by the device itself should not be validated
against common display standards. Most validations can be
circumvented by not passing ranges to validate against.
This is not true for the dutry cycle.
While a 10 percent duty cycle seems low for a CRT device
(which should not be below 20 percent) it may too high for
modern display panels.