Mainly reverted commit 0a546480a5c2e0c66f746a91cb0b5f592c06b357.
This mixed up save & restore.
Save is used for saving GPU state when *entering* the Xserver, Restore for
restoring the old state when leaving. That includes the cursor (there could be
a cursor active on VTs - though unlikely).
Upon entering we save state, then *reload* the cursor image we just *know* to
be correct. No need for saving that on LeaveVT.