Getting PLL limits from AtomBIOS.
AtomBIOS has information about the PLL_IN min/max VCO min/max and
pixel clock max frequency.
As the VCO_min frequency in AtomBIOS is 0 in most BIOSes that have
been checked the code uses a fixed value of 600MHz for now.

Also the names of the variables have been modified to match the
nomenclature of atombios.h better.