Change CombineStrings() to AppendString().
The semantics have changed. Now the first string is deallocated
in some sense following realloc(). When the first element is NULL
a new string will be created (like with strdup()). When the second
string is NULL the first one will be returned unaltered. The
same will happen in an out of memory condition.