Make sure that config.h is included in every file that includes rhd.h.
rhd.h contains structures that are different wether
libpciaccess is used or not.
Therefore we need to include config.h before rhd.h.
To avoid double inclusions of headers we don't include
it in rhd.h itself but check there if the define for
it is set. If not we simply error on build.