Be smarter about the monitor section by selectively ignoring it.
Too many users are hitting this horribly outdated paradigm of one driver, one
device, one monitor. Most of the configuration utilities out there still
insist on creating a monitor section inside the config file, with a
completely useless statement which enables DPMS which happens to be enabled
per default anyway and cannot be disabled (the latter because of broken
option handling in X dpms code).

This hurts us because, when a handcreated monitor section is there, a sane
driver should assume that this configured monitor is just there to add further
restrictions to already known monitors, or to make up for lack of autodetected
monitor information. This is why a driver should use this information.

But apparently, the only real-world option is to ignore this section.
Sanity doesn't count.

So, new handling:
* complain about monitor section if autodetected info is present.
* use configured monitor if no autodetected info is present, fall back
  to default timing if no configured monitor is there.
* option "useconfiguredmonitor" is there to enable the so desiring user
  to make the configured monitor matter in the autodetected case.

Personally, i believe this to be overly complicated handling, and wrong.
It's the third millenium, we can have multiple monitors already.