Don't shut down outputs that have active connectors attached.
Something that cannot be solved with the current RandR model is that multiple
connectors might be attached to the same output. They cannot be driven with
different crtcs then, but RandR sees them as different outputs because of a
missing abstraction layer. This implementation will silently set one crtc or
the other, and maybe even won't warn.