[PATCH] Avoid double -l when linking rhd_conntest
If libpci is found via pkg-config, PCIUTILS_LIBS will contain
something like "-L/path/to/lib -lpci -lz". So we get rid of the
explicit -lpci and -lz in this case.

In case pkg-config does not find libpci, though, we just revert
to the previous behaviour of assuming that "-lpci -lz" will work.

In any case, we make sure linking works before actually using it.