Modes: Add DPI calculation.
This adds code to provide useful DPI calculation. It follows a server argument,
or uses the newly introduced option forcedpi, or tries to average out the DPI
of the various monitors.

This is of course a stopgap solution, as no DPI calculation can ever be valid
throughout the whole server life, as it varies with varying resolutions and
varying monitors. But at least it is better than having nothing at all.