Restore: Fix VGA textmode restore when VSYNC length is 0.
When using the VGA emulation support, crtc VSYNC can be set to null, and the
VGA timing will still apply. But when you save and then restore this value,
things become very tricky, as the 0 length VSYNC provides bad sync timing and
displays will then go into power saving.

By setting the VSYNC to an acceptable value for a shortish time, the VGA engine
apparently is able to latch in correctly, and as such the issue is seemingly
fixed.

I have no idea about a better solution for this issue, preferably one involving
the VGA support itself, but this, although ugly, does seem to work.

Denise H. G is very much resposible for fixing this issue, as she pointed out
that a null value in VSYNC is the cause of the problem here.