Fix some issues when reading PCI ROM on POSTed systems.
This stuff is rather magic.
The SPI ROM interface needs to be set up. The SCK period
needs to be set to be increased (power up default should be
save).
The GPIO pins seem to control the SPI ROM, so the GPIO pins
must be set to be driven by hardware.