Change heuristics to check for an unPOSTed card.
The code used to look for the location and size of the VRAM
scratch area in the BIOS to determine if the device has been
POSTed or not. This value reads 0 on unPOSTed BIOSes.
This however is not true for all cards.
Now we simply look where the BIOS has originated.