Fix blaning problem in bug #13853.
RHDSaveScreen() is a DIX function. It gets called right from the DPMS
extension. Non of the DIX function know anything about screen switching. Thus
the protection of the hardware against accidental switches needs to come from
the DDX (which includes the driver - if it provies functions which may be
called from DIX directly).
Adding a check for vtSema seems to fix the problem.