Fully implement the SaveScreen() function.
The argument to SaveScreen() is not simply on/off.
xf86IsUnblank() needs to be called to actually match the
DIX argument to on/off. Also the input timer needs to be
reset on unblank. Furthermore the ScreenPtr may be NULL
under certain circumstances thus it also needs to be checked
before dereferencing.