Changing R6XX_MC_VM_MISC_OFFSET to R6XX_HDP_NONSURFACE_BASE.
HDP_NONSURFACE_BASE is the official ATI name.
Originally we found this register in code tables called by AsicInit.
It was given the name MC_VM_MISC_OFFSET because we didn't know the
official name of it as we never received docs on it from ATI.

Now it appeared in the 'radeon' driver under it's official name.