Restructure FMT block handling.
The FMT block is directly linked to the CRTC block but only the ditital
outputs seem to be connected to it. The FMT block preformats the pixel
data for the display doing things like dithering which will give the
impression of a higher color depth on 18bit digital displays.
Since the FMT block is connected to the CRTC we save/restore it in the
rhd_crtc.c while we program it from within DigMode().