Add support for scaled video modes.
Here mode validation will be done differently depending if we validate
an unscaled or a mode that is to be scaled.
From the mode that is to be scaled to a different display resolution we
only need the viewport size, thus we only need to validate it against the
scaler.
Finding the true display mode to validate against is a little trickier:
we look for the monitor that is to be attached and pick the preferred
mode.