Add syntesized CVT modes to mode list for scaling.
For RandR we need to synthesize modes for scaling as there is no way
to obtain the full mode pool.
We take a list of well known display resolutions and generate CVT
mode lines from them. This is mainly done so that those modes can
pass validation as we only need HDisplay and VDisplay from these.