Attempt to fail modes if scaled mode is invalid.
Unfortunately it has no effect if we fail a mode in rhdRROutputModeFixup().
Instead we need to validate this mode earlier (before rhdRROutputModeValid()
is called and add it to the rhdRandrOutputRec. From there we can pick it
to create the timing mode in rhdRROutputModeFixup() and pass it as an
additional argument to RHDRRModeFixup().