Add a per monitor native mode.
The native mode should be the mode that coincides with the
native resolution of a monitor. For an LCD this would be
the number of horizontal and vertical pixel.
On panels we assume it's either the higest resolution in
an EDID block or the panel mode supplied by AtomBIOS.

For any other monitor that can do DDC it's the first detailed
timing if the monitor has the preferred mode flag bit set.