Move mode validation after scaler validation.
This groups all validations that deal with actual timing parameters
together. These validations affect the mode whose timing parameters are
used ie the one that is set for the display.
All validations which take place before scaler validation deal with
fb parameters, thus deal with the mode which we scale from.
In the situation where we use scaling the display mode is validated
once and before any other mode list validation takes place.
When the mode list is validated only the modes from which we scale
from are only validated against fb and the scaler as those modes have
no influence on the actual timing.