Fix heuristic for RR mode validation to find out about scaled mode.
Also always attach scaledTo mode to rhd CRTC struct so we don't need to
pass it around. The scaledTo mode is only used when the crtc specific
scaling validation is available. Thus we can attach it there.

Move collection of ScaledToMode to rhdRROutputGetModes(). This way the
scaled to mode will be redetermined whenever a the output is reprobed.