Restructured OutputPrivates + numerous bug fixes.
The output privates contain AtomBIOS device information.
Which device to pick for a specific output isn't known however until we
have probed the output.
Also the object headers don't maintain the devices in the same order
as the outputs in the src/dest table. We therefore need a heuristic
to pick the correct device for an output.
Along this numerous other bugs were fixed.