ACCEL: Move variable to track 3D engine state to rhdAccel and rename.
This variable describes the state of the 3D engine (if it is set up in
the context of a DRI client or in the context of the Xserver).
Thus it should live in the rhdAccel accel state structure.
Also the name has been changed to XHas3DEngineState to be more
self explanatory.