DRI: Adjust files to build against the drm kernel headers.
The radeon_* files used header files with the same content
as the kernel drm headers. This patch adds the necessary modifications
to make those files build against the kernel header files.
Since this driver can run with older API versions and it cannot
be expected that all users have the latest drm header files
installed it was necessary to copy radeon_drm.h from the kernel.
Eventually this may also have to be done with sarea.h which however
is hopefully less likely to change.
This also does away with the insanity that different files built
against different header files introduced with patch
2f12947ba7a9d14760d229005b7cfc2f90bf022b.