2D: Changed test for AccelMethod to test for presence of method specific data struct.
Testing for the presence of the data struct is more portable. It requires
that it is made shure the data struct is only there when the accel method
is really initialized.
This patch also changees the (historic) name accel to xaa to clearer
reflect that this data structure belong to XAA.