AtomBIOS: Make sure the OutputDriverPrivate is always set.
To disable outputs we must make sure the output driver privates are
always set, even when the output is not connected. This is done in the
RandR detection function.
The problem: AtomBIOS sense needs to know the device type to sense -
which may be different from the device type currently set in the output
private if the output can be used on more than one connector (like DACs).

To overcome this we ask for the device type matching the connector in
the sense function itself and discard it afterwards.
If one output type was sensed on outputs that may have more than one
output type we make sure to not sense again or overwrite the driver
private with a different device type.

For outouts that don't have anything connected we just leave the
type that was set last.

We really need to find the device type that's set at startup time.
We may be able to do so using the BIOS scratch registers.
However in this case we'd have to set the device type a lot later.