Crtc: Call CRTC Save/Restore wrapper directly.
These functions used to save/restore CRTC registers directly.
Since the save/restore functions of the different CRTC subsystems
have been separated out it only serves as a wrapper.
Therefore it doesn't need to live in the Crtc structure any
more but can be moved up to a generic level that is called
directly.