AtomBIOS/Output: Fix backlight control handling for panels.
AtomBIOS controlled panel backlight brightness control should work with this
patch, now - where supported by AtomBIOS. There seem to be a lot of laptops
where this is not the case. I'm not sure how to detect this without looking
at the AtomBIOS code. Could BIOS scratch register 2 bit 8-16 == 0 be an
indication?