MC: Tune MC on RV515 and RS690 to avoid glitches on hi res modes.
To avoid glitches on Hi-Res modes due to linebuffer starvation
the MC can get tuned to make linebuffer requests urgent when a
timer has expired before the linebuffer has again obtained access
to the framebuffer.
This mechanism eists in this form in RV515 and derivatives
as well as on RS690.
Later chips implement a different mechanism.
This initial implementation sets the timer to the lowest
value pussible so the linebuffer becomes priority very early.
An algorithm to estimate the correct value is on the TODO list.