CS EXA: Add download from screen.
We claim an indirect buffer all to ourselves to serve as the location
to which the engine can blit to, from which we can have a relatively
cheap memcpy to the destination buffer. We need to have a nasty idle
before this though, and it might pay to handle this more intelligently.