I2C: Improve probing for monitor presence.
The ASUS M2A-VM (RS690) motherboard comes in two variants: with and without
HDMI. The HDMI variant ships with a riser card to be plugged into the PEG
slot. Apart from this riser card both boards seem to be identical.
When no riser card is installed all I2C slaves on the I2C line assigned to
the HDMI connector send an ACK.
Therefore to scheck fro the presence DDC it's not sufficient to look for
the ACK but also to read the first two bytes and check if they belong to a
DDC header.