AtomBIOS: Temporary fix for CRTC Restore with AtomBIOS.
EnableCRTCMemReq controls memory access per CRTC. It however
accesses registers in the DC blocks which are shared between
the CRTCs.
Thus if we don't want to save these registers separately
we need to restore CRTCs in the opposite order we have saved
them in.
The real fix would be to save/restore the affected registers
separately.