AtomBIOS: Check all lists if a register has already been saved.
This is the proper fix to the problem 'hot-fixed' in
         f5b0d7c452fd49e8581485b5a8efc7a0ad37ba9f.
It we record all save lists (ie. lists from all subsystems) that have been
used and check them all if a register has already been saved.
This saves us from saving a register twice (the second time with an already
modified value) in which case restore becomes very order dependent.