DDC: Make RHD_DDC_* proper records into an array of I2C register tables.
This way we can find out which GPIO pin is used for the SDA and SCL
lines and set this up appropriately. R5 and R6 always use DDC line
pairs. RS690 and and DCE3.0 chipsets can use lines from different
pairs.
This fixes the handling of GPIO lines properly for RS690. So far
it has only worked on RS690 purely by accident.