RandR: Improve handling physical outputs which map to multiple RandR outputs.
Outputs like DACs can drive multiple RandR outputs (one at a time).
RandR doesn't seem to have a notion of this. Thus when probing for
one RandR outputs we need to probe all other possible RandR outputs
which map to this same physical output and make a decision which one
we report to RandR as connected.
For now and for the case of the DACs we just report TVout as disconnected
this needs to be change once TV output is fixed.