VGA: Calculate offset from base address correctly.
The MC_FB_LOCATION (on R6xx and higher the MC_VM_FB_LOCATION)
mark the base of the range that's send to the local framebuffer controller.
This (instead of the HDP_FB_LOCATION) should be used for offset calculation.
A call is added to the MC subsystem to query this value in an ASIC specific
way.