Fix README update on FreeBSD (sed substitution)
FreeBSD sed does not understand \t or \n in the substitute
string.

This commit fixes the updating of README as it only uses \t
which can easily be replaced by a tab character.

The man/radeonhd.man update uses \n which is more difficult
to handle and thus will be handled in a later commit.