Fix and improve RhdParseBooleanOption.
This commit fixes a small bug in the new RhdParseBooleanOption function,
which was causing a xserver crash with the coherent option set.
It also improves the matching capability of this function by adding an
match all case and a new return value indicating that a given name
wasn't found in the option.