DRI/CS: Set drmFD to -1 after closing the device.
The CS merge destroyed the heuristic by which we detect if DRI is active
or not by introducing its own heuristic. CS itself requests the fd of the
drm device and checks if it is < 0. Unfortunately drmFD is neither
initialized to -1 nor reset after closing the device.
Adresses 2nd issue in:
https://bugs.freedesktop.org/show_bug.cgi?id=18474