RandR: Improve heuristics to determine of an output is connected.
On a multi output connector DDC and HPD are per connector - not per
output.
Thus they don't allow to destinguish which of the outputs is connected.
If the output provides a sense function it can be used to determine the
connection state. If it doesn't - but the other output(s) on the same
connector do, they can be probed applying the inverse logic: if another
output is connected, this one most likely isn't.

Code for this existed already for outputs which provide HPD. We now
also use it for outputs which don't.