MC/Idle: ASSERT() the right setup order, don't change an unidled MC.
AllIdle() needs to be called before MCSetup() to make sure the
MC is idle. We can only idle the MC when the VGA subsystem is
disabled which happens in VGADisable() which therefore needs
to be called first.
If we are not idle in MCSetup() we fail without setting up the MC.
If MCSetup() fails we bail. If the MC is not idle during restore
we don't restore the MC.

Conflicts: