r6xx/r7xx EXA: Optimize overlapping copy
When source and destination blocks are only offset horizontally, it
appears to be unnecessary to perform careful, segment-by-segment copy.
The code path that does this is taken out completely.

For the case where offset is only vertical, copying is now done by
height of the non-overlapping area each time, instead of always
line-by-line.