Add Fallback DAC Load Detection Method for chips >= RV620.
There are two methods of doing DAC load detection on DCE3.x chips: one
works 'on the fly' ie doesn't produce noticeable flickering but
doesn't work as reliably on some chips RS780 especially in connection
with a KVM switch where it may produce a false negative, the other may
produce some visible artefacts but also seems to work reliably with KVM
switches. It's the one used in most AtomBIOSes.
We now use both methods: if the 'on the fly' one produces a negative
result we fall back to the other one and let it decide wether something
is connected or not.