I2C: Fix I2C readout.
The Radeon driver attempts to save/restore the hw scaler registers
by reading/writing the range of registers which address the video scaler.
While doing so it also touches the TAP Control index/data registers.
Uncontrolled touching of the data register seems to have an adverse
effect on the ability to run HW supported I2C read out. This running
RadeonHD after running the Radeon driver results in no DDC data.
The reason for this has not been understood however it has been found
experimentally that reading back the scaler data register is able to
'fix' this issue.
We put this hacky workaround in the code until we obtain a better
understanding of the issue.