Probe: Only allow this driver when KMS is not present.
If kernel modesetting (KMS) is detected and active, we abort the
PciProbe and do not touch the GPU in any way. Then at least the KMS
console continues working, and there is an error message for the user
to read.
This is only temporary until KMS has been added to this driver.